+++ Controllers/PersonController.cs
@@ -1,5 +1,7 @@
 ﻿using Library.Models;
 using Library.Services.PersonService;
+using Library.Services.SessionService;
+using Library.Services.UserService;
 using Library.Utils;
 using Microsoft.AspNetCore.Mvc;
 
@@ -14,9 +16,12 @@ namespace Library.Controllers {
 
         private readonly IPersonService _personService;
 
+        private readonly ISessionService _sessionService;
 
-        public PersonController(IPersonService personService) {
+
+        public PersonController(IPersonService personService, ISessionService sessionService) {
             _personService = personService;
+            _sessionService = sessionService;
         }
 
 
@@ -26,6 +31,12 @@ namespace Library.Controllers {
         /// <returns>Página de manutenção de pessoas.</returns>
         [HttpGet]
         public async Task<IActionResult> Manage() {
+
+            if (!_sessionService.IsTheSessionActive()) {
+                return RedirectToAction("Login", "Login");
+            }
+
+            _sessionService.SetLayout(this);
 
             var personsResp = await _personService.GetPersons();
 
@@ -47,8 +58,16 @@ namespace Library.Controllers {
         /// </summary>
         /// <returns>Página para cadastro de uma nova pessoa.</returns>
         [HttpGet]
-        public IActionResult Register() {
+        public async Task<IActionResult> Register() {
+
+            if (!_sessionService.IsTheSessionActive()) {
+                return RedirectToAction("Login", "Login");
+            }
+
+            _sessionService.SetLayout(this);
+
             return View();
+
         }
 
 
@@ -60,19 +79,23 @@ namespace Library.Controllers {
         [HttpPost]
         public async Task<IActionResult> Register(PersonModel person) {
 
+            if (!_sessionService.IsTheSessionActive()) {
+                return RedirectToAction("Login", "Login");
+            }
+
             if (ModelState.IsValid) {
 
                 var registerPersonResp = await _personService.RegisterPerson(person);
 
                 if (registerPersonResp.Successful) {
 
-                    TempData[Constants.success_message] = registerPersonResp.Message;
+                    TempData[Constants.SUCCESS_MESSAGE] = registerPersonResp.Message;
 
                     return RedirectToAction("Manage");
 
                 } else {
 
-                    TempData[Constants.error_message] = registerPersonResp.Message;
+                    TempData[Constants.ERROR_MESSAGE] = registerPersonResp.Message;
 
                     return View(person);
 
@@ -80,7 +103,7 @@ namespace Library.Controllers {
 
             } else {
 
-                TempData[Constants.error_message] = "Dados da pessoa incorretos!";
+                TempData[Constants.ERROR_MESSAGE] = "Dados da pessoa incorretos!";
 
                 return View(person);
 
@@ -96,6 +119,12 @@ namespace Library.Controllers {
         /// <returns>Página para edição de uma pessoa.</returns>
         [HttpGet]
         public async Task<IActionResult> Edit(Guid id) {
+
+            if (!_sessionService.IsTheSessionActive()) {
+                return RedirectToAction("Login", "Login");
+            }
+
+            _sessionService.SetLayout(this);
 
             var personResp = await _personService.GetPerson(id);
 
@@ -124,19 +153,23 @@ namespace Library.Controllers {
         [HttpPost]
         public async Task<IActionResult> Edit(PersonModel person) {
 
+            if (!_sessionService.IsTheSessionActive()) {
+                return RedirectToAction("Login", "Login");
+            }
+
             if (ModelState.IsValid) {
 
                 var editPersonResp = await _personService.EditPerson(person);
 
                 if (editPersonResp.Successful) {
 
-                    TempData[Constants.success_message] = editPersonResp.Message;
+                    TempData[Constants.SUCCESS_MESSAGE] = editPersonResp.Message;
 
                     return RedirectToAction("Manage");
 
                 } else {
 
-                    TempData[Constants.error_message] = editPersonResp.Message;
+                    TempData[Constants.ERROR_MESSAGE] = editPersonResp.Message;
 
                     return View(person);
 
@@ -144,7 +177,7 @@ namespace Library.Controllers {
 
             } else {
 
-                TempData[Constants.error_message] = "Dados da pessoa incorretos!";
+                TempData[Constants.ERROR_MESSAGE] = "Dados da pessoa incorretos!";
 
                 return View(person);
 
@@ -160,6 +193,12 @@ namespace Library.Controllers {
         /// <returns>Página para a exclusão de uma pessoa.</returns>
         [HttpGet]
         public async Task<IActionResult> Delete(Guid id) {
+
+            if (!_sessionService.IsTheSessionActive()) {
+                return RedirectToAction("Login", "Login");
+            }
+
+            _sessionService.SetLayout(this);
 
             var personResp = await _personService.GetPerson(id);
 
@@ -188,19 +227,23 @@ namespace Library.Controllers {
         [HttpPost]
         public async Task<IActionResult> Delete(PersonModel person) {
 
+            if (!_sessionService.IsTheSessionActive()) {
+                return RedirectToAction("Login", "Login");
+            }
+
             if (ModelState.IsValid) {
 
                 var deletePersonResp = await _personService.DeletePerson(person);
 
                 if (deletePersonResp.Successful) {
 
-                    TempData[Constants.success_message] = deletePersonResp.Message;
+                    TempData[Constants.SUCCESS_MESSAGE] = deletePersonResp.Message;
 
                     return RedirectToAction("Manage");
 
                 } else {
 
-                    TempData[Constants.error_message] = deletePersonResp.Message;
+                    TempData[Constants.ERROR_MESSAGE] = deletePersonResp.Message;
 
                     return View(person);
 
@@ -208,7 +251,7 @@ namespace Library.Controllers {
 
             } else {
 
-                TempData[Constants.success_message] = "Dados da pessoa incorretos!";
+                TempData[Constants.SUCCESS_MESSAGE] = "Dados da pessoa incorretos!";
 
                 return View(person);
